+++ include/aikido/control/KinematicSimulationJointCommandExecutor.hpp
@@ -24,13 +24,28 @@ public:
   explicit KinematicSimulationJointCommandExecutor(
       ::dart::dynamics::MetaSkeletonPtr metaskeleton);
 
+  /// DoF-only Constructor
+  ///
+  /// \param dofs DoFs to execute commands on.
+  explicit KinematicSimulationJointCommandExecutor(
+      std::vector<::dart::dynamics::DegreeOfFreedom*> dofs);
+
   virtual ~KinematicSimulationJointCommandExecutor();
 
   // Documentation inherited
   virtual std::future<int> execute(
       const std::vector<double>& command,
+      const std::chrono::duration<double>& timeout,
+      const std::chrono::system_clock::time_point& timepoint) override;
+
+  // Documentation inherited
+  virtual std::future<int> execute(
+      const std::vector<double>& command,
       const std::chrono::duration<double>& timeout
-      = std::chrono::duration<double>(1)) override;
+      = std::chrono::duration<double>(1.0)) override
+  {
+    return this->execute(command, timeout, std::chrono::system_clock::now());
+  }
 
   /// \copydoc Executor::step()
   ///
@@ -73,7 +88,9 @@ using KinematicSimulationPositionExecutor
     = KinematicSimulationJointCommandExecutor<ExecutorType::POSITION>;
 using KinematicSimulationVelocityExecutor
     = KinematicSimulationJointCommandExecutor<ExecutorType::VELOCITY>;
-// Note: No effort simulation without dynamics
+// Note: No effort simulation without dynamics, treat like velocity.
+using KinematicSimulationEffortExecutor
+    = KinematicSimulationJointCommandExecutor<ExecutorType::EFFORT>;
 
 } // namespace control
 } // namespace aikido