+++ CHANGELOG.md
@@ -1,11 +1,14 @@
 ## Aikido 0 (prerelease)
 
-### 0.5.0 (2021-XX-XX)
+### 0.5.0 (2022-XX-XX)
 
 * Control
 
   * Fixed bug in RosTrajectoryExecutor [#596](https://github.com/personalrobotics/aikido/pull/596)
   * Added generic and position/velocity/effort Executor types [#602](https://github.com/personalrobotics/aikido/pull/602)
+  * Added JacobianExecutor wrapping any Velocity/EffortExecutor [#605](https://github.com/personalrobotics/aikido/pull/605)
+  * Implemented generic Executors into Robot class [#614](https://github.com/personalrobotics/aikido/pull/614)
+  * Added VisualServoingVelocityExecutor wrapping JacobianExecutor [#620](https://github.com/personalrobotics/aikido/pull/620)
 
 * IO
 
@@ -17,8 +20,9 @@
 
 * RViz
 
-  * Fix addTSRMarker Bug: [#618](https://github.com/personalrobotics/aikido/pull/618)
   * Propogate removal of WorldInteractiveMarkerViewer to Aikidopy: [#599](https://github.com/personalrobotics/aikido/pull/599)
+  * Fix addTSRMarker Bug: [#618](https://github.com/personalrobotics/aikido/pull/618)
+  * Implement clear markers in InteractiveMarkerViewer: [#625](https://github.com/personalrobotics/aikido/pull/625)
 
 * State Space